requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
Django==5.2.1
django-celery-beat==2.8.1
django-timezone-field==7.1
djangorestframework==3.16.0
drf-yasg==1.21.10
inflection==0.5.1
kombu==5.5.4
packaging==25.0
pillow==11.2.1
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.2.13
django-cors-headers==4.9.0
